requirements.txt
blinker==1.6.2
cachelib==0.10.2
click==8.1.7
Flask==2.3.3
Flask-Admin==1.6.1
Flask-Login==0.6.2
Flask-SQLAlchemy==3.0.5
Flask-WTF==1.2.1
greenlet==2.0.2
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
psycopg2==2.9.7
python-dotenv==1.0.0
SQLAlchemy==2.0.20
typing_extensions==4.7.1
Werkzeug==2.3.7
WTForms==3.0.1
